Created test_types_protocol.py
Moved protocol types tests to new file, removed from __init__ collections
tests file
